refactor: extract custom hooks from bundle-browser, icon-browser, and icon-cart
- Extract useBundleBrowser hook from bundle-browser.tsx (21 useState + effects)
- Extract useIconBrowser hook from metallic-icon-browser.tsx (21 useState + effects)
- Extract useIconCart hook from icon-cart.tsx (12 useState + effects)
- Remove 2 eslint-disable comments for react-hooks/exhaustive-deps by
  using a fetchIconsRef pattern to avoid stale closures without adding
  fetchIcons to effect dependency arrays
- Each hook lives co-located as use-[name].ts next to its component
- All existing behavior preserved; lint, typecheck, and build pass clean
